distrohub-backend/requirements.txt
fastapi[standard]==0.128.0
psycopg[binary]==3.3.2
pydantic==2.12.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.21
asyncpg==0.31.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
supabase==2.27.0
uvicorn[standard]==0.30.0
httpx==0.27.0

